[CUDAExtension] Refactor the CUDA architecture flag retrieval function
